Refactor: 쿠키 캘린더 activeMonth 수정
